refactor(WinningReferee): checkWin 역할 분리
